requirements.txt
langchain
langchain-groq
pandas 
streamlit
python-dotenv
